fix(localize): folder restructure and Chrome corrections
