+++ packages/localize/src/date/utils/trim.js
