+++ packages/localize/src/number/utils/normalSpaces.js
