+++ packages/localize/src/number/utils/normalize-format-number-to-parts/forceNormalSpaces.js
@@ -1,9 +1,9 @@
-import { normalSpaces } from './normalSpaces.js';
+import { normalSpaces } from '../normalSpaces.js';
 
 /**
  * Parts with forced "normal" spaces
  *
- * @typedef {import('../../types/LocalizeMixinTypes').FormatNumberPart} FormatNumberPart
+ * @typedef {import('../../../../types/LocalizeMixinTypes').FormatNumberPart} FormatNumberPart
  * @param {FormatNumberPart[]} formattedParts
  * @returns {FormatNumberPart[]}
  */
