+++ packages/localize/src/number/utils/normalize-get-currency-name/forceCurrencyNameForPHPEnGB.js
@@ -1,0 +1,10 @@
+/**
+ * @param {string} currencyName
+ */
+export function forceCurrencyNameForPHPEnGB(currencyName) {
+  if (currencyName === 'Philippine pesos') {
+    // eslint-disable-next-line no-param-reassign
+    currencyName = 'Philippine pisos';
+  }
+  return currencyName;
+}
